[ALPS09780515] ASoC: Exception occurs when killing the audio server
[Description]
An exception with the message "AFE memif enable before set_addr" occurs
when the audio server (HAL) is repeatedly killed, which indicates an
improper sequence in enabling the Audio Front-End memory interface
before setting its address.

[Cause]
The memory interface is not being properly disabled when the audio
server terminated unexpectedly.

[Solution]
The `mtk_afe_fe_hw_free()` will be invoked when the audio server
terminates. Therefore we handle the disabling of memory interfaces in
this function as well.

[Test]
1. Connect USB headset
2. Make a phone call (SIM card required)
3. Execute `adb root`
4. Execute `adb shell pkill audioserver`
5. Repeat Step.4 and ensue no errors occur

MTK-Commit-Id: b12480bea433266296fab6535c6fdba6546746e1

Signed-off-by: Shawn Sung <shawn.sung@mediatek.com>
CR-Id: ALPS09780515
Feature: alsa_sound_soc
Change-Id: Ic6d9a27a53241ca90b154b00e7886dd43f424359